requirements.txt
python-dotenv
nextcord
requests
python-socketio[asyncio_client]
aiohttp